rework some c++ syntax to be a bit more backwards compatible for slightly older gcc versions available on raspberry pi also, cleanup the system target a bit to be more compat with older make

+++ model/value.cpp
@@ -157,7 +157,7 @@ Value::Value(bool b)
   _bool = b;
 }
 
-Value::Value(size_t n)
+Value::Value(uint64_t n)
 {
   _id = LUA_TNUMBER;
   _type = "number";
